requirements.txt
selenium>=3.14
pytest-selenium
Appium-Python-Client
pytest